Fixed some warnings

+++ src/chunk.cxx
@@ -105,16 +105,16 @@ void Chunk::deserializeSection(std::int8_t nthSection, const nbt::Int8Vector& bl
 
 void bestTerrainGenerator(Chunk& ch)
 {
-  for (std::size_t x = 0; x < Chunk::ChunkMaxX; ++x)
-    for (std::size_t z = 0; z < Chunk::ChunkMaxZ; ++z)
+  for (std::int32_t x = 0; x < Chunk::ChunkMaxX; ++x)
+    for (std::int32_t z = 0; z < Chunk::ChunkMaxZ; ++z)
       ch(x, 0, z) = BlockType::Bedrock;
   
-  for (std::size_t x = 0; x < Chunk::ChunkMaxX; ++x)
-    for (std::size_t z = 0; z < Chunk::ChunkMaxZ; ++z)
+  for (std::int32_t x = 0; x < Chunk::ChunkMaxX; ++x)
+    for (std::int32_t z = 0; z < Chunk::ChunkMaxZ; ++z)
       ch(x, 1, z) = BlockType::Glowstone;
   
-  for (std::size_t x = 0; x < Chunk::ChunkMaxX; ++x)
-    for (std::size_t z = 0; z < Chunk::ChunkMaxZ; ++z)
+  for (std::int32_t x = 0; x < Chunk::ChunkMaxX; ++x)
+    for (std::int32_t z = 0; z < Chunk::ChunkMaxZ; ++z)
       ch(x, 2, z) = BlockType::Leaves2;
 }
